Menampilkan data

+++ database/migrations/2023_05_13_011103_create_employees_table.php
@@ -17,7 +17,7 @@ class CreateEmployeesTable extends Migration
             $table->id();
             $table->string('nama');
             $table->enum('jeniskelamin', ['laki-laki', 'perempuan']);
-            $table->integer('notelpon');
+            $table->bigInteger('notelpon');
             $table->timestamps();
         });
     }
